requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
chromadb==0.4.18
openai==1.3.7
pymupdf>=1.24.0
python-multipart==0.0.6
pydantic==2.5.0
tiktoken==0.5.2
sqlalchemy==2.0.23
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

